move projection chart in different file
